Aggregate habits by day Move default habit list to data module Add habit duration

+++ feature/summary/src/main/java/com/githukudenis/summary/ui/SummaryUiEvent.kt
@@ -3,5 +3,5 @@ package com.githukudenis.summary.ui
 sealed interface SummaryUiEvent {
     object Refresh : SummaryUiEvent
     data class ShowError(val error: UserError) : SummaryUiEvent
-    data class CheckHabit(val habitId: Int) : SummaryUiEvent
+    data class CheckHabit(val habitId: Long) : SummaryUiEvent
 }
